Move AggregationMethod and FilterOperation to core
